Revert "std.os.linux: fix kernel_timespec.nsec to be isize"
This reverts commit af88310caf86b1b1e84d9f0764af522b402241b2.

I conflated __kernel_timespec with __kernel_old_timespec. i64 is correct.

+++ lib/std/os/linux.zig
@@ -9438,7 +9438,7 @@ pub const timezone = extern struct {
 /// The timespec struct used by the kernel.
 pub const kernel_timespec = extern struct {
     sec: i64,
-    nsec: isize,
+    nsec: i64,
 
     /// For use with `utimensat` and `futimens`.
     pub const NOW: timespec = .{
